# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Crewquarters Control API
  description: 'Control plane for Crewquarters. JSON is camelCase. Errors use `{error: {code, message,
    requestId, details}}`. State-changing requests need the session cookie, an allowed Origin, and the
    `X-CSRF-Token` header. Mutating requests accept an `Idempotency-Key` header.'
  version: 0.1.0
paths:
...
  /api/v1/schedules/preview:
    post:
      tags:
      - schedules
      summary: Preview the next occurrences of a cron expression in a timezone
      operationId: preview_schedule_api_v1_schedules_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePreviewIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulePreviewOut'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/schedules:
    post:
      tags:
      - schedules
      summary: Create a schedule
      operationId: create_schedule_api_v1_schedules_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreateIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    get:
      tags:
      - schedules
      summary: List schedules, newest first (sort by nextRunAt client-side for 'upcoming')
      operationId: list_schedules_api_v1_schedules_get
      parameters:
      - name: installationId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Installationid
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_ScheduleOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/schedules/{schedule_id}:
    get:
      tags:
      - schedules
      summary: Get a schedule
      operationId: get_schedule_api_v1_schedules__schedule_id__get
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Schedule Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    patch:
      tags:
      - schedules
      summary: Update a schedule
      operationId: update_schedule_api_v1_schedules__schedule_id__patch
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Schedule Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePatchIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOut'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
    delete:
      tags:
      - schedules
      summary: Delete a schedule
      operationId: delete_schedule_api_v1_schedules__schedule_id__delete
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Schedule Id
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
...
components:
  schemas:
...
    ErrorDetail:
      properties:
        code:
          type: string
          title: Code
          examples:
          - MODEL_CAPACITY_EXCEEDED
        message:
          type: string
          title: Message
        requestId:
          anyOf:
          - type: string
          - type: 'null'
          title: Requestid
        details:
          additionalProperties: true
          type: object
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OccurrenceOut:
      properties:
        at:
          type: string
          format: date-time
          title: At
          description: Occurrence instant in UTC.
        local:
          type: string
          title: Local
          description: Local wall-clock time with UTC offset, ISO 8601.
        zoneAbbreviation:
          type: string
          title: Zoneabbreviation
      type: object
      required:
      - at
      - local
      - zoneAbbreviation
      title: OccurrenceOut
...
    Page_ScheduleOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ScheduleOut'
          type: array
          title: Items
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
          description: Opaque cursor for the next page; null on the last page.
      type: object
      required:
      - items
      title: Page[ScheduleOut]
...
    ReadinessCheck:
      properties:
        name:
          type: string
          enum:
          - enabled
          - permissions
          - configuration
          - architecture
          - connection
          - model
          title: Name
        status:
          type: string
          enum:
          - ok
          - missing
          - needs_attention
          title: Status
        detail:
          type: string
          title: Detail
        resource:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource
      type: object
      required:
      - name
      - status
      - detail
      title: ReadinessCheck
...
    ScheduleCreateIn:
      properties:
        installationId:
          type: string
          format: uuid
          title: Installationid
        cron:
          type: string
          title: Cron
          examples:
          - 0 10 * * *
        timezone:
          type: string
          title: Timezone
          description: IANA timezone name.
          examples:
          - Asia/Kolkata
        misfirePolicy:
          type: string
          enum:
          - fire_once
          - skip
          title: Misfirepolicy
          default: fire_once
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - installationId
      - cron
      - timezone
      title: ScheduleCreateIn
    ScheduleOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        installationId:
          type: string
          format: uuid
          title: Installationid
        agentName:
          type: string
          title: Agentname
        ready:
          type: boolean
          title: Ready
          description: Whether the agent's model, connections, and config are ready now.
        blockers:
          items:
            $ref: '#/components/schemas/ReadinessCheck'
          type: array
          title: Blockers
          description: Readiness checks that are not ok.
        cron:
          type: string
          title: Cron
        timezone:
          type: string
          title: Timezone
        misfirePolicy:
          type: string
          enum:
          - fire_once
          - skip
          title: Misfirepolicy
        enabled:
          type: boolean
          title: Enabled
        nextRunAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Nextrunat
        nextOccurrences:
          items:
            $ref: '#/components/schemas/OccurrenceOut'
          type: array
          title: Nextoccurrences
        lastFiredAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lastfiredat
        lastRunId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Lastrunid
        version:
          type: integer
          title: Version
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - installationId
      - agentName
      - ready
      - blockers
      - cron
      - timezone
      - misfirePolicy
      - enabled
      - nextRunAt
      - nextOccurrences
      - lastFiredAt
      - lastRunId
      - version
      - createdAt
      - updatedAt
      title: ScheduleOut
    SchedulePatchIn:
      properties:
        version:
          type: integer
          title: Version
        cron:
          anyOf:
          - type: string
          - type: 'null'
          title: Cron
        timezone:
          anyOf:
          - type: string
          - type: 'null'
          title: Timezone
        misfirePolicy:
          anyOf:
          - type: string
            enum:
            - fire_once
            - skip
          - type: 'null'
          title: Misfirepolicy
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      type: object
      required:
      - version
      title: SchedulePatchIn
    SchedulePreviewIn:
      properties:
        cron:
          type: string
          title: Cron
        timezone:
          type: string
          title: Timezone
        count:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Count
          default: 3
      type: object
      required:
      - cron
      - timezone
      title: SchedulePreviewIn
    SchedulePreviewOut:
      properties:
        cron:
          type: string
          title: Cron
        timezone:
          type: string
          title: Timezone
        occurrences:
          items:
            $ref: '#/components/schemas/OccurrenceOut'
          type: array
          title: Occurrences
      type: object
      required:
      - cron
      - timezone
      - occurrences
      title: SchedulePreviewOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: cq_session
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
    service:
      type: http
      scheme: bearer
      description: Internal service credential
servers:
- url: /
  description: Same origin as the web UI